Inline the stop-kills-children fixture so pytest doesn't report SKIPPED
The previous test_spawns_child.py used pytest.mark.skipif gated on an env
var so it would skip when collected by the regular suite. That left a
"SKIPPED (invoked only by the runner stop test)" line in every pytest run.

Move the fixture source into a string constant in the integration test
and write it to tmp_path at runtime. The fixture is no longer a real
test file in tests/, so pytest never collects it and there is nothing
to skip.
