+++ tests/test_pytest_runner/test_pytest_runner_stop_kills_children.py
@@ -10,6 +10,22 @@ from pytest_fly.pytest_runner import PytestRunner
 
 from ..paths import get_temp_dir
 
+_FIXTURE_TEST_SOURCE = """\
+import os
+import subprocess
+import sys
+import time
+
+
+def test_spawns_child():
+    pid_file = os.environ["PYTEST_FLY_CHILD_PID_FILE"]
+    child = subprocess.Popen([sys.executable, "-c", "import time; time.sleep(120)"])
+    with open(pid_file, "w") as f:
+        f.write(str(child.pid))
+    time.sleep(120)
+    assert True
+"""
+
 
 def _wait_pid_gone(pid: int, timeout: float = 10.0) -> bool:
     deadline = time.time() + timeout
@@ -26,11 +42,14 @@ def test_pytest_runner_stop_kills_children(app, tmp_path):
     data_dir = get_temp_dir(test_name)
     run_guid = generate_uuid()
 
+    # Write the fixture test to tmp_path so it's never collected by the regular
+    # suite (a real test_*.py file would always show up as SKIPPED).
+    fixture_file = Path(tmp_path, "test_spawns_child.py")
+    fixture_file.write_text(_FIXTURE_TEST_SOURCE)
     pid_file = Path(tmp_path, "child.pid")
-    os.environ["PYTEST_FLY_SPAWN_CHILD_TEST"] = "1"
     os.environ["PYTEST_FLY_CHILD_PID_FILE"] = str(pid_file)
     try:
-        scheduled_tests = [ScheduledTest(node_id="tests/test_spawns_child.py", singleton=False, duration=None, coverage=None)]
+        scheduled_tests = [ScheduledTest(node_id=str(fixture_file), singleton=False, duration=None, coverage=None)]
 
         runner = PytestRunner(run_guid, scheduled_tests, number_of_processes=1, data_dir=data_dir, update_rate=1.0)
         runner.start()
@@ -48,5 +67,4 @@ def test_pytest_runner_stop_kills_children(app, tmp_path):
 
         assert _wait_pid_gone(child_pid), f"orphaned child pid {child_pid} still alive after runner.stop()"
     finally:
-        os.environ.pop("PYTEST_FLY_SPAWN_CHILD_TEST", None)
         os.environ.pop("PYTEST_FLY_CHILD_PID_FILE", None)
